書き込みの表示はログインにかかわらず共通なのでクラスの継承
