+++ app/Http/Controllers/Dashboard/ThreadsController.php
@@ -2,34 +2,12 @@
 
 namespace App\Http\Controllers\Dashboard;
 
-use App\Http\Controllers\Controller;
-use App\Http\Controllers\Dashboard\NotLoggedIn\ThreadsController as NotLoggedInThreadsController;
+use App\Http\Controllers\Dashboard\NotLoggedIn\ThreadsController as Controller;
 use App\Models\Hub;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 
 class ThreadsController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
     /**
      * Store a newly created resource in storage.
      *
@@ -81,69 +59,5 @@ class ThreadsController extends Controller
         }
 
         (new ThreadImagePathController)->store($request, $message_id);
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Support\Collection | void
-     */
-    public function show(Request $request)
-    {
-        if (!Auth::check()) {
-            return (new NotLoggedInThreadsController)->show($request->thread_id, $request->max_message_id);
-        }
-
-        $thread = Hub::where('id', '=', $request->thread_id)
-            ->where('is_enabled', '=', 1)
-            ->first();
-        switch ($thread->thread_category_type) {
-            case '学科':
-                return (new DepartmentThreadController)->show($request->user()->email, $request->thread_id, $request->max_message_id);
-            case '学年':
-                return (new CollegeYearThreadController)->show($request->user()->email, $request->thread_id, $request->max_message_id);
-            case '部活':
-                return (new ClubThreadController)->show($request->user()->email, $request->thread_id, $request->max_message_id);
-            case '授業':
-                return (new LectureThreadController)->show($request->user()->email, $request->thread_id, $request->max_message_id);
-            case '就職':
-                return (new JobHuntingThreadController)->show($request->user()->email, $request->thread_id, $request->max_message_id);
-            default:
-                return null;
-        }
-    }
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        //
     }
 }
